move apply_template to helpers
